Add some management functions to desktop list
This commit changes some of the interactive functions to first look if
the desktop list is being shown and, if so, if there is a desktop
specified under the cursor. If either of these conditions is not true,
they fall back to asking the user for a name, as before.